Add animation support
The `generate` function was moved to `generate_image` and another
function called `generate_animation` was added. This function generates
and animation using matplotlib's animation module (and hence supports
most of the things that `animation` supports). The interpolation to
perform the animation is a simple linear interpolation right now, but it
looks fine, especially in the polar coordinate system. May add
new/better/more efficient methods later on.