FuncAnimation, random state, function generator
This commit contains multiple additions:
1. The animation is now produced using a FuncAnimation object instead of
ArtistAnimation object. This reduces the memory consumption
significantly as now I don't store every state of the artist in a list
but use a generator to generate it at every frame.
2. Support to set a random seed or a random state (as supported by the
numpy.random.default_rng().__setstate__()). This random state is picked
up from a path to a json file that would have been stored by some
previous run.
3. As suggested in the above point, the state/configuration is stored as
a json file in the same directory as the image/video.
4. Added support for random generation of a function. So, if a user does
not provide a function, the library will generate a random function by
creating a random expression tree using various unary and binary
functions from numpy and use these functions to generate the images.
(Experimental) The string representation of these generated function is
also stored in the configuration json.